chore(review): 한 줄 리뷰 store로 모두 이동
