+++ packages/entity/src/review.ts
@@ -1,55 +1,8 @@
 import type { BaseRequestData } from './appMetadata';
 
-export interface OneLineReview {
-  reviewUuid: string;
-  storeId: number;
-  nickname: string;
-  profileImage: string;
-  content: string;
-  rating: number;
-  createdAt: string;
-  images: string[];
-}
-
 export interface ReviewTag {
   id: number;
   name: string;
-}
-
-export interface OnelineReviewRequests {
-  userUuid: string;
-  content: string;
-  rating: number;
-}
-export interface CreateOnelineReviewRequestFormData {
-  storeUuid: string;
-  request: OnelineReviewRequests;
-  // images?: string[];
-  images?: File[];
-}
-
-export interface CreateOnelineReviewResponse {
-  reviewUuid: string;
-  storeId: number;
-  content: string;
-  rating: number;
-  createdAt: string;
-  images: string[];
-}
-
-export interface StoreOnelineReivewRequest {
-  storeUuid: string;
-}
-
-export interface StoreOnelineReivewData {
-  reviewUuid: string;
-  storeId: number;
-  nickname: string;
-  profileImage: string;
-  content: string;
-  rating: number;
-  createdAt: string;
-  images: string[];
 }
 
 export interface ReviewUpdateData {
@@ -57,16 +10,6 @@ export interface ReviewUpdateData {
 }
 
 export interface ReviewRepository {
-  getStoreOnelineReviews(
-    data: BaseRequestData<StoreOnelineReivewRequest>,
-  ): Promise<StoreOnelineReivewData[]>;
-
-  createOnelineReview({
-    data,
-  }: BaseRequestData<CreateOnelineReviewRequestFormData>): Promise<
-    CreateOnelineReviewResponse[]
-  >;
-
   // 내가 쓴 리뷰 조회
   getMine(data: BaseRequestData<unknown>): Promise<unknown>;
